IntegraçãoParcialFront2Back Professor e feed logado separados entre servidor e client, a parte do servidor faz requisições para o back e essas requisições mandam os objetos; No caso professores, para os compenentes específicos receberem a requisição [feedLogMain.tsx e professorloggedmain.tsx]
