+++ grupo2brabo/src/app/components/Nav.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 const Nav: React.FC = () => {
   return (
     <div className="w-screen h-24 bg-[#A4FED3] flex items-center justify-between pr-4 pl-4">
-    <img src="caminho_para_a_logo_unb.png" className="h-16" alt="Logo" />
+    <img src="/caminho_para_a_logo_unb.png" className="h-16" alt="Logo" />
 </div>
   );
 }
