+++ grupo2brabo/src/app/components/interfacesPadrao.tsx
@@ -1,0 +1,7 @@
+export interface Professor {
+    name: string
+    id:number
+    departamento: string
+    foto: string
+    date: string
+}
